+++ src/main/java/com/pradip/sewearn/dto/receive/SewEarnReceiveDto.java
@@ -16,7 +16,7 @@ public class SewEarnReceiveDto {
 
     private Integer totalReceivedQuantity;
 
-    private Double totalEarning;
+    private Long totalEarning;
 
     private Boolean markAsCompleted;
