+++ src/main/java/com/pradip/sewearn/projection/DailyEarningProjection.java
@@ -1,8 +1,10 @@
 package com.pradip.sewearn.projection;
 
+import org.hibernate.mapping.TableOwner;
+
 import java.time.LocalDate;
 
 public interface DailyEarningProjection {
     LocalDate getCompletedDate();
-    Double getTotalEarning();
+    Long getTotalEarning();
 }
