+++ src/main/java/com/pradip/sewearn/projection/MaterialEarningProjection.java
@@ -2,5 +2,5 @@ package com.pradip.sewearn.projection;
 
 public interface MaterialEarningProjection {
     String getMaterialName();
-    Double getTotalEarning();
+    Long getTotalEarning();
 }
